# backend/docs/postgres-reference/openapi.design.yaml
openapi: 3.0.3
info:
  title: School Admission Management API
  version: 1.0.0
  description: |
    REST API for enquiries, applications, workflow transitions, assessments,
    payments and reporting.

    **Conventions**
    - Base path `/api/v1`. JSON only. Timestamps are ISO-8601 UTC.
    - Auth: OAuth2 / OIDC bearer JWT. Guardians sign in with OTP; the JWT carries a `roles` claim.
    - Errors use RFC 9457 `application/problem+json`.
    - Mutating calls accept an `Idempotency-Key` header; payment calls require it.
    - Applications use optimistic locking: send `If-Match: <version>` on writes; a stale value returns `409`.
    - Lists are cursor-paginated: `?limit=25&cursor=...`.
servers:
  - url: https://api.example-school.in/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/otp/request:
    post:
      tags: [Auth]
      summary: Send a one-time password to a guardian's phone or email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, destination]
              properties:
                channel: { type: string, enum: [SMS, WHATSAPP, EMAIL] }
                destination: { type: string, example: '+919812345678' }
      responses:
        '202': { description: OTP sent (always 202 to avoid account enumeration) }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Exchange an OTP for tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination, code]
              properties:
                destination: { type: string }
                code: { type: string, minLength: 6, maxLength: 6 }
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
                  expiresIn: { type: integer, example: 900 }
        '401': { $ref: '#/components/responses/Problem' }

  # ------------------------------------------------------------------- Leads
  /leads:
    get:
      tags: [Leads]
      summary: Search leads
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Name or phone }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/LeadStatus' } }
        - { name: source, in: query, schema: { type: string } }
        - { name: assignedTo, in: query, schema: { type: string, format: uuid } }
        - { name: followUpBefore, in: query, schema: { type: string, format: date-time }, description: Overdue follow-ups }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Page of leads
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Lead' }
    post:
      tags: [Leads]
      summary: Create a lead manually (walk-in, phone, fair)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnquiryCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lead' }
  /leads/{leadId}/assign:
    post:
      tags: [Leads]
      summary: Assign a lead to a counsellor
      parameters: [{ $ref: '#/components/parameters/LeadId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
      responses:
        '200': { description: Assigned }
  /leads/{leadId}/followups:
    post:
      tags: [Leads]
      summary: Schedule or log a follow-up
      parameters: [{ $ref: '#/components/parameters/LeadId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dueAt]
              properties:
                dueAt: { type: string, format: date-time }
                note: { type: string }
      responses:
        '201': { description: Created }
  /leads/{leadId}/convert:
    post:
      tags: [Leads]
      summary: Convert a lead into an application draft
      parameters: [{ $ref: '#/components/parameters/LeadId' }]
      responses:
        '201':
          description: Application created at the first workflow stage
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }

  # ------------------------------------------------------------ Applications
...
  /applications/{applicationId}/invoices:
    get:
      tags: [Payments]
      summary: Invoices for an application
      parameters: [{ $ref: '#/components/parameters/ApplicationId' }]
      responses:
        '200':
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Invoice' }
  /invoices/{invoiceId}/payments:
    post:
      tags: [Payments]
      summary: Create a gateway order for an invoice
      description: Requires `Idempotency-Key`. The client completes payment with the returned gateway data; the stage changes only after the webhook confirms.
      parameters:
        - { name: invoiceId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gateway: { type: string, example: razorpay }
                method: { type: string, enum: [UPI, CARD, NETBANKING] }
      responses:
        '201':
          description: Gateway order
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentId: { type: string, format: uuid }
                  gateway: { type: string }
                  gatewayOrderId: { type: string }
                  amount: { type: number }
                  currency: { type: string, example: INR }
        '409': { description: Invoice already paid }
...
  /reports/funnel:
    get:
      tags: [Reports]
      summary: Stage funnel with conversion
      parameters:
        - { name: academicYear, in: query, required: true, schema: { type: string } }
        - { name: classCode, in: query, schema: { type: string } }
        - { name: campusId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Reached-count per stage
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    stage: { type: string }
                    reached: { type: integer }
                    conversionFromEnquiry: { type: number, example: 0.42 }
  /reports/sources:
    get:
      tags: [Reports]
      summary: Enquiry-to-admission conversion by source
      parameters:
        - { name: academicYear, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: Per-source stats
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    source: { type: string }
                    enquiries: { type: integer }
                    applications: { type: integer }
                    admitted: { type: integer }
...
  /reports/seats:
    get:
      tags: [Reports]
      summary: Seat availability per class
      parameters:
        - { name: academicYear, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: Seats
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    classCode: { type: string }
                    seatsTotal: { type: integer }
                    seatsUsed: { type: integer }
                    seatsLeft: { type: integer }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string, format: uuid }
    LeadId:
      name: leadId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        traceId: { type: string }
    Page:
      type: object
      properties:
        items: { type: array, items: {} }
        nextCursor: { type: string, nullable: true }
    LeadStatus:
      type: string
      enum: [NEW, CONTACTED, INTERESTED, VISIT_SCHEDULED, CONVERTED, LOST]
    EnquiryCreate:
      type: object
      required: [guardianName, phone, source]
      properties:
        guardianName: { type: string }
        phone: { type: string, example: '+919812345678' }
        email: { type: string, format: email }
        childName: { type: string }
        childDob: { type: string, format: date }
        classCode: { type: string }
        source: { type: string, example: WEBSITE }
        utm:
          type: object
          additionalProperties: { type: string }
        consent:
          type: boolean
          description: Guardian consent to be contacted (required under DPDP)
    Lead:
      type: object
      properties:
        id: { type: string, format: uuid }
        guardianName: { type: string }
        phone: { type: string }
        childName: { type: string }
        classCode: { type: string }
        source: { type: string }
        status: { $ref: '#/components/schemas/LeadStatus' }
        assignedTo: { type: string, format: uuid, nullable: true }
        nextFollowUpAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    Applicant:
      type: object
      required: [fullName, dob]
      properties:
        fullName: { type: string }
        dob: { type: string, format: date }
        gender: { type: string }
        previousSchool: { type: string }
...
    Application:
      type: object
      properties:
        id: { type: string, format: uuid }
        applicationNo: { type: string, example: APP27-000123 }
        classCode: { type: string }
        stage: { type: string, example: APPLICATION }
        version: { type: integer, format: int64 }
        admissionNo: { type: string, nullable: true }
        applicant: { $ref: '#/components/schemas/Applicant' }
        assignedTo: { type: string, format: uuid, nullable: true }
        submittedAt: { type: string, format: date-time, nullable: true }
...
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        invoiceNo: { type: string }
        amount: { type: number }
        currency: { type: string }
        status: { type: string, enum: [ISSUED, PARTIAL, PAID, CANCELLED, EXPIRED] }
        dueAt: { type: string, format: date-time }
...